requirements.txt
faiss-cpu==1.10.0
fastapi==0.115.12
google-auth==2.38.0
google-cloud-storage==2.19.0
numpy==2.2.4
openai==1.72.0
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyMuPDF==1.25.5
requests==2.32.3
sentence-transformers==4.0.2
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
uvicorn==0.34.0
pymongo[srv]==4.12.0
huggingface_hub[hf_xet]